HR {survey}	R Documentation

Wrappers for specifying PPS designs

Description

The Horvitz-Thompson estimator and the Hartley-Rao approximation require information in addition to the sampling probabilities for sampled individuals. These functions allow this information to be supplied.

Usage

HR(psum=NULL, strata = NULL)
ppsmat(jointprob, tolerance = 1e-04) 

Arguments

psum 	 The sum of squared sampling probabilities for the population, divided by the sample size, as a single number or as a vector for stratified sampling

strata 	 Stratum labels, of the same length as psum, if psum is a vector

jointprob 	 Matrix of pairwise sampling probabilities for the sampled individuals

tolerance 	 Tolerance for deciding that the covariance of sampling indicators is zero

Value

An object of class HR or ppsmat, suitable for supplying as the pps argument to svydesign.

See Also

election for examples of PPS designs

Examples

HR(0.1)

[Package survey version 3.18 Index]