bootweights {survey}	R Documentation

Compute survey bootstrap weights

Description

Bootstrap weights for infinite populations are created by sampling with replacement from the PSUs in each stratum. For finite populations an empirical population is created as described by Canty and Davision (1999). The bootstrap is primarily useful when most strata have many PSUs.

This bootstrap is strictly appropriate only when the first stage of sampling is a simple or stratified random sample of PSUs with or without replacement, and not (eg) for PPS sampling.

Usage

bootweights(strata, psu, replicates = 50, fpc = NULL,
         fpctype = c("population", "fraction", "correction"),
         compress = TRUE)

Arguments

strata 	 Identifier for sampling strata (top level)

psu 	 Identifier for primary sampling units

replicates 	 Number of bootstrap replicates

fpc 	 Finite population correction

fpctype 	 Is fpc the population size, sampling fraction, or 1-sampling fraction?

compress 	 Should the replicate weights be compressed?

Value

A set of replicate weights

References

Canty AJ, Davison AC. (1999) Resampling-based variance estimation for labour force surveys. The Statistician 48:379-391

See Also

as.svrepdesign

Examples

[Package survey version 3.1-1 Index]