compressWeights {survey}	R Documentation

Compress replicate weight matrix

Description

Many replicate weight matrices have redundant rows, such as when weights are the same for all observations in a PSU. This function produces a compressed form. Methods for as.matrix and as.vector extract and expand the weights.

Usage

compressWeights(rw, ...)
## S3 method for class 'svyrep.design':
compressWeights(rw,...)
## S3 method for class 'repweights_compressed':
as.matrix(x,...)
## S3 method for class 'repweights_compressed':
as.vector(x,...)

Arguments

rw 	 A set of replicate weights or a svyrep.design object

x 	 A compressed set of replicate weights

... 	 For future expansion

Value

An object of class repweights_compressed or a svyrep.design object with repweights element of class repweights_compressed

See Also

jknweights,as.svrepdesign

Examples

data(api)
dclus1<-svydesign(id=~dnum, weights=~pw, data=apiclus1, fpc=~fpc)
rclus1c<-as.svrepdesign(dclus1,compress=TRUE)
rclus1<-as.svrepdesign(dclus1,compress=FALSE)

[Package survey version 3.1-1 Index]