Association between leprosy and BCG vaccination

These data are in a paper by JNK Rao and colleagues, on score tests for complex survey data. External information (not further specified) suggests the functional form for the Age variable.

data("myco")

Format

A data frame with 516 observations on the following 6 variables.

Age

Age in years at the midpoint of six age strata

Scar

Presence of a BCG vaccination scar

n

Sampled number of cases (and thus controls) in the age stratum

Ncontrol

Number of non-cases in the population

wt

Sampling weight

leprosy

case status 0/1

Details

The data are a simulated stratified case-control study drawn from a population study conducted in a region of Malawi (Clayton and Hills, 1993, Table 18.1). The goal was to examine whether BCG vaccination against tuberculosis protects against leprosy (the causative agents are both species of _Mycobacterium_). Rao et al have a typographical error: the number of non-cases in the population in the 25-30 age stratum is given as 4981 but 5981 matches both the computational output and the data as given by Clayton and Hills.

Source

JNK Rao, AJ Scott, and C Rao, J., Scott, A., & Skinner, C. (1998). QUASI-SCORE TESTS WITH SURVEY DATA. Statistica Sinica, 8(4), 1059-1070.

Clayton, D., & Hills, M. (1993). Statistical Models in Epidemiology. OUP

Examples

data(myco)
dmyco<-svydesign(id=~1, strata=~interaction(Age,leprosy),weights=~wt,data=myco)

m_full<-svyglm(leprosy~I((Age+7.5)^-2)+Scar, family=quasibinomial, design=dmyco)
m_age<-svyglm(leprosy~I((Age+7.5)^-2), family=quasibinomial, design=dmyco)
anova(m_full,m_age)
#> Working (Rao-Scott+F) LRT for Scar
#>  in svyglm(formula = leprosy ~ I((Age + 7.5)^-2) + Scar, design = dmyco, 
#>     family = quasibinomial)
#> Working 2logLR =  10.47836 p= 0.0013711 
#> df=1;  denominator df= 502

## unweighted model does not match
m_full
#> Stratified Independent Sampling design (with replacement)
#> svydesign(id = ~1, strata = ~interaction(Age, leprosy), weights = ~wt, 
#>     data = myco)
#> 
#> Call:  svyglm(formula = leprosy ~ I((Age + 7.5)^-2) + Scar, design = dmyco, 
#>     family = quasibinomial)
#> 
#> Coefficients:
#>       (Intercept)  I((Age + 7.5)^-2)               Scar  
#>           -4.4777          -370.5582            -0.5479  
#> 
#> Degrees of Freedom: 515 Total (i.e. Null);  502 Residual
#> Null Deviance:	    27.99 
#> Residual Deviance: 27.24 	AIC: NA
glm(leprosy~I((Age+7.5)^-2)+Scar, family=binomial, data=myco)
#> 
#> Call:  glm(formula = leprosy ~ I((Age + 7.5)^-2) + Scar, family = binomial, 
#>     data = myco)
#> 
#> Coefficients:
#>       (Intercept)  I((Age + 7.5)^-2)               Scar  
#>            0.1126            48.7230            -0.4392  
#> 
#> Degrees of Freedom: 515 Total (i.e. Null);  513 Residual
#> Null Deviance:	    715.3 
#> Residual Deviance: 709.6 	AIC: 715.6