Open and close DBI connections

A database-backed survey design object contains a connection to a database. This connection will be broken if the object is saved and reloaded, and the connection should ideally be closed with close before quitting R (although it doesn't matter for SQLite connections). The connection can be reopened with open.

# S3 method for DBIsvydesign
open(con, ...)
# S3 method for DBIsvydesign
close(con, ...)

Arguments

con

Object of class DBIsvydesign

...

Other options, to be passed to dbConnect or dbDisconnect.

Value

The same survey design object with the connection opened or closed.

See also

Examples

if (FALSE) {
library(RSQLite)
dbclus1<-svydesign(id=~dnum, weights=~pw, fpc=~fpc,
data="apiclus1",dbtype="SQLite",
dbname=system.file("api.db",package="survey"))

dbclus1
close(dbclus1)
dbclus1
try(svymean(~api00, dbclus1))

dbclus1<-open(dbclus1)
open(dbclus1)
svymean(~api00, dbclus1)
}