Compute variance from replicates

Compute an appropriately scaled empirical variance estimate from replicates. The mse argument specifies whether the sums of squares should be centered at the point estimate (mse=TRUE) or the mean of the replicates. It is usually taken from the mse component of the design object.

svrVar(thetas, scale, rscales, na.action=getOption("na.action"), 
  mse=getOption("survey.replicates.mse"),coef)

Arguments

thetas

matrix whose rows are replicates (or a vector of replicates)

scale

Overall scaling factor

rscales

Scaling factor for each squared deviation

na.action

How to handle replicates where the statistic could not be estimated

mse

if TRUE, center at the point estimated, if FALSE center at the mean of the replicates

coef

The point estimate, required only if mse==TRUE

Value

covariance matrix.

See also