svy.varcoef {survey}	R Documentation

Sandwich variance estimator for glms

Description

Computes the sandwich variance estimator for a generalised linear model fitted to data from a complex sample survey. Designed to be used internally by svyglm.

Usage

svy.varcoef(glm.object, design,std.errors = c("linearized",
                 "Bell-McCaffrey", "Bell-McCaffrey-2"), degf = FALSE)

Arguments

glm.object
A glm object

design
A survey.design object

std.errors
The kind of standard errors to compute

degf
Whether to compute the adjusted degrees of freedom along with Bell-McCaffrey standard errors

Value

A variance matrix

Author(s)

Thomas Lumley

See Also

svyglm,svydesign, svyCprod

[Package survey version 4.5 Index]