svygofchisq {survey}	R Documentation

Test of fit to known probabilities

Description

A Rao-Scott-type version of the chi-squared test for goodness of fit to prespecified proportions. The test statistic is the chi-squared statistic applied to the estimated population table, and the reference distribution is a Satterthwaite approximation: the test statistic divided by the estimated scale is compared to a chi-squared distribution with the estimated df.

Usage

svygofchisq(formula, p, design, ...)

Arguments

formula
Formula specifying a single factor variable

p
Vector of probabilities for the categories of the factor, in the correct order (will be rescaled to sum to 1)

design
Survey design object

...
Other arguments to pass to svytotal, such as na.rm

Value

An object of class htest

See Also

chisq.test, svychisq, pchisqsum

Examples

data(api)
dclus2<-svydesign(id=~dnum+snum, fpc=~fpc1+fpc2, data=apiclus2)

true_p <- table(apipop$stype)

svygofchisq(~stype,dclus2,p=true_p)
svygofchisq(~stype,dclus2,p=c(1/3,1/3,1/3))

[Package survey version 4.1-1 Index]