Two-stage least-squares for instrumental variable regression

Estimates regressions with endogenous covariates using two-stage least squares. The function uses ivreg from the AER package for the main computations, and follows the syntax of that function.

svyivreg(formula, design, ...)

Arguments

formula

formula specification(s) of the regression relationship and the instruments. See Details for details

design

A survey design object

...

For future expansion

Details

Regressors and instruments for svyivreg are specified in a formula with two parts on the right-hand side, e.g., y ~ x1 + x2 | z1 + z2 + z3, where x1 and x2 are the regressors and z1, z2, and z3 are the instruments. Note that exogenous regressors have to be included as instruments for themselves. For example, if there is one exogenous regressor ex and one endogenous regressor en with instrument in, the appropriate formula would be y ~ ex + en | ex + in. Equivalently, this can be specified as y ~ ex + en | . - en + in, i.e., by providing an update formula with a . in the second part of the formula.

Value

An object of class svyivreg

References

https://notstatschat.rbind.io/2019/07/16/adding-new-functions-to-the-survey-package/

See also