bootweights {survey}	R Documentation

Compute survey bootstrap weights

Description

Bootstrap weights for infinite populations are created by sampling with replacement from the PSUs in each stratum. subbootweights() samples n-1 PSUs from the n available (Rao and Wu), bootweights samples n (Canty and Davison).

bootweights() is most useful when the strata are large and may have large sampling fractions, subbootweights() is useful when the strata have a small number of PSUs and the sampling fractions are small.

This bootstrap is strictly appropriate only when the first stage of sampling is a simple or stratified random sample of PSUs with or without replacement, and not (eg) for PPS sampling.

Usage

bootweights(strata, psu, replicates = 50, fpc = NULL,
         fpctype = c("population", "fraction", "correction"),
         compress = TRUE)
subbootweights(strata, psu, replicates = 50,
         compress = TRUE)

Arguments

strata 	 Identifier for sampling strata (top level)

psu 	 Identifier for primary sampling units

replicates 	 Number of bootstrap replicates

fpc 	 Finite population correction

fpctype 	 Is fpc the population size, sampling fraction, or 1-sampling fraction?

compress 	 Should the replicate weights be compressed?

Value

A set of replicate weights

References

Canty AJ, Davison AC. (1999) Resampling-based variance estimation for labour force surveys. The Statistician 48:379-391

Rao JNK, Wu CFJ. Bootstrap inference for sample surveys. Proc Section on Survey Research Methodology. 1993 (866–871)

See Also

as.svrepdesign

[Package survey version 3.18 Index]