Confidence intervals for regression parameters

Computes confidence intervals for regression parameters in svyglm objects. The default is a Wald-type confidence interval, adding and subtracting a multiple of the standard error. The method="likelihood" is an interval based on inverting the Rao-Scott likelihood ratio test. That is, it is an interval where the working model deviance is lower than the threshold for the Rao-Scott test at the specified level.

# S3 method for svyglm
confint(object, parm, level = 0.95, method = c("Wald", "likelihood"), ddf = NULL, ...)

Arguments

object

svyglm object

parm

numeric or character vector indicating which parameters to construct intervals for.

level

desired coverage

method

See description above

ddf

Denominator degrees of freedom for "likelihood" method, to use a t distribution rather than norma. If NULL, use object$df.residual

...

for future expansion

Value

A matrix of confidence intervals

References

J. N. K. Rao and Alistair J. Scott (1984) On Chi-squared Tests For Multiway Contigency Tables with Proportions Estimated From Survey Data. Annals of Statistics 12:46-60

See also

Examples

data(api)
dclus2<-svydesign(id=~dnum+snum, fpc=~fpc1+fpc2, data=apiclus2)

m<-svyglm(I(comp.imp=="Yes")~stype*emer+ell, design=dclus2, family=quasibinomial)
confint(m)
#>                   2.5 %     97.5 %
#> (Intercept)  0.56476264 3.70319468
#> stypeH      -4.28558148 0.93393621
#> stypeM      -3.70649097 1.34532080
#> emer        -0.08588860 0.04659558
#> ell         -0.05693166 0.02448049
#> stypeH:emer -0.13853776 0.07244824
#> stypeM:emer -0.20697834 0.14040185
confint(m, method="like",ddf=NULL, parm=c("ell","emer"))
#>            2.5 %     97.5 %
#> ell  -0.05891077 0.02469912
#> emer -0.08637239 0.04994329
#> attr(,"levels")
#> [1] 0.95