regTermTest {survey}	R Documentation

Wald test for a term in a regression model

Description

Provides Wald test and working likelihood ratio (Rao-Scott) test of the hypothesis that all coefficients associated with a particular regression term are zero (or have some other specified values). Particularly useful as a substitute for anova when not fitting by maximum likelihood. The Wald tests use a chisquared or F distribution, the LRT uses a linear combination of chisquared distributions as in pchisqsum.

Usage

regTermTest(model, test.terms, null=NULL,df=Inf, method=c("Wald","LRT"))

Arguments

model 	 A model object with coef and vcov methods

test.terms 	 Character string or one-sided formula giving name of term or terms to test

null 	 Null hypothesis values for parameters. Default is zeros

df 	 Denominator degrees of freedom for an F test. If NULL these are estimated from the model

method 	 If "Wald", the Wald-type test; if "LRT" the Rao-Scott test based on the estimated log likelihood ratio

Value

An object of class regTermTest or regTermTestLRT.

References

Rao, JNK, Scott, AJ (1984) "On Chi-squared Tests For Multiway Contingency Tables with Proportions Estimated From Survey Data" Annals of Statistics 12:46-60.

See Also

anova, vcov, contrasts,pchisqsum

Examples

 data(esoph)
 model1 <- glm(cbind(ncases, ncontrols) ~ agegp + tobgp * 
     alcgp, data = esoph, family = binomial())
 anova(model1)

 regTermTest(model1,"tobgp")
 regTermTest(model1,"tobgp:alcgp")
 regTermTest(model1, ~alcgp+tobgp:alcgp)

 data(api)
 dclus2<-svydesign(id=~dnum+snum, weights=~pw, data=apiclus2)
 model2<-svyglm(I(sch.wide=="Yes")~ell+meals+mobility, design=dclus2, family=quasibinomial())
 regTermTest(model2, ~ell)
 regTermTest(model2, ~ell,df=NULL)
 regTermTest(model2, ~ell, method="LRT")
 regTermTest(model2, ~ell+meals, method="LRT")

[Package survey version 3.18 Index]