svrVar {survey}	R Documentation

Compute variance from replicates

Description

Compute an appropriately scaled empirical variance estimate from replicates.

Usage

svrVar(thetas, scale, rscales, na.action=getOption("na.action"))

Arguments

thetas 	 matrix whose rows are replicates (or a vector of replicates)

scale 	 Overall scaling factor

rscales 	 Scaling factor for each squared deviation

na.action 	 How to handle replicates where the statistic could not be estimated

Value

covariance matrix.

See Also

svrepdesign, as.svrepdesign, brrweights, jk1weights, jknweights

[Package survey version 3.18 Index]