svy.varcoef {survey}	R Documentation

Sandwich variance estimator for glms

Description

Computes the sandwich variance estimator for a generalised linear model fitted to data from a complex sample survey. Designed to be used internally by svyglm.

Usage

svy.varcoef(glm.object, design)

Arguments

glm.object 	 A glm object

design 	 A \survey.design object

Value

A variance matrix

Author(s)

Thomas Lumley

See Also

svyglm,svydesign, svyCprod

Examples

[Package survey version 3.1-1 Index]