svyglm {survey}	R Documentation

Survey-weighted generalised linear models.

Description

Fit a generalised linear model to data from a complex survey design, with inverse-probability weighting and design-based standard errors.

Usage

## S3 method for class 'survey.design':
svyglm(formula, design, subset=NULL, ...)
## S3 method for class 'svyrep.design':
svyglm(formula, design, subset=NULL, ..., rho=NULL,
return.replicates=FALSE, na.action)
## S3 method for class 'svyglm':
summary(object, correlation = FALSE,  ...) 

Arguments

formula 	 Model formula

design 	 Survey design from svydesign or svrepdesign. Must contain all variables in the formula

subset 	 Expression to select a subpopulation

... 	 Other arguments passed to glm or summary.glm

rho 	 For replicate BRR designs, to specify the parameter for Fay's variance method, giving weights of rho and 2-rho

return.replicates 	 Return the replicates as a component of the result?

object 	 A svyglm object

correlation 	 Include the correlation matrix of parameters?

na.action 	 Handling of NAs

Details

There is no anova method for svyglm as the models are not fitted by maximum likelihood. The function regTermTest may be useful for testing sets of regression terms.

Value

An object of class svyglm.

Author(s)

Thomas Lumley

See Also

svydesign, svrepdesign,as.svrepdesign, glm, regTermTest

Examples

  data(api)

  glm(api00~ell+meals+mobility, data=apipop)

  dstrat<-svydesign(id=~1,strata=~stype, weights=~pw, data=apistrat, fpc=~fpc)
  dclus2<-svydesign(id=~dnum+snum, weights=~pw, data=apiclus2)
  rstrat<-as.svrepdesign(dstrat)
  rclus2<-as.svrepdesign(dclus2)

  summary(svyglm(api00~ell+meals+mobility, design=dstrat))
  summary(svyglm(api00~ell+meals+mobility, design=dclus2))
  summary(svyglm(api00~ell+meals+mobility, design=rstrat))
  summary(svyglm(api00~ell+meals+mobility, design=rclus2))

  ## use quasibinomial, quasipoisson to avoid warning messages
  summary(svyglm(sch.wide~ell+meals+mobility, design=dstrat, family=quasibinomial()))

 

[Package survey version 3.1-1 Index]