svyolr {survey}	R Documentation

Proportional odds and related models

Description

Fits cumulative link models: proportional odds, probit, complementary log-log, and cauchit.

Usage

svyolr(formula, design, ...)
## S3 method for class 'survey.design2'
svyolr(formula, design, start, ..., na.action = na.omit, method = c("logistic", 
    "probit", "cloglog", "cauchit"))
## S3 method for class 'svyrep.design'
svyolr(formula,design,...,return.replicates=FALSE, 
    multicore=getOption("survey.multicore"))

Arguments

formula
Formula: the response must be a factor with at least three levels

design
survey design object

...
dots

start
Optional starting values for optimization

na.action
handling of missing values

multicore
Use multicore package to distribute computation of replicates across multiple processors?

method
Link function

return.replicates
return the individual replicate-weight estimates

Value

An object of class svyolr

Author(s)

The code is based closely on polr() from the MASS package of Venables and Ripley.

See Also

svyglm, regTermTest

Examples

data(api)
dclus1<-svydesign(id=~dnum, weights=~pw, data=apiclus1, fpc=~fpc)
dclus1<-update(dclus1, mealcat=cut(meals,c(0,25,50,75,100)))

m<-svyolr(mealcat~avg.ed+mobility+stype, design=dclus1)
m

## Use regTermTest for testing multiple parameters
regTermTest(m, ~avg.ed+stype, method="LRT")

[Package survey version 4.1-1 Index]